api/requirements.txt
chromadb>=0.5.23
langchain>=0.3.14
pydantic>=2.10.4
beautifulsoup4>=4.12.2
langgraph>=0.2.61
requests>=2.31.0
langchain_openai>=0.3.3
langchain_community>=0.3.16
langchain-chroma>=0.2.1
tweepy>=4.15.0
fastapi>=0.115.7
datasets>=3.2.0
pyyaml>=6.0.1
pillow>=10.2.0
transformers>=4.40.1
uvicorn>=0.34.0
openai>=1.59.4